requirements.txt
# Core dependencies
gymnasium[box2d]>=0.29.0
numpy>=1.24.0
torch>=2.0.0
tensorboard>=2.14.0

# Training utilities
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and experiment tracking
wandb>=0.15.0  # Optional: for advanced experiment tracking

# Video recording
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

tabulate
